build: separate edit-images from userInfo
