+++ packages/web/src/components/AvatarUserForm/index.js
@@ -1,0 +1,1 @@
+export { default } from './AvatarUserForm';
