+++ packages/web/src/components/MainInfoForm/index.js
@@ -1,0 +1,1 @@
+export { default } from './MainInfoForm';
